pyproject.toml
[tool.poetry]
name = "mammopy"
version = "0.0.15"
description = "A unified solution for mammogram image analysis and interpretation"
authors = ["UEF Cancer <uef.cancergroup@gmail.com>"]
readme = "README.md"
include = ["segmentation_models/*", "segmentation_models/base/*", "segmentation_models/encoders/*", "segmentation_models/unet/*", "segmentation_models/utils/*"]


[tool.poetry.dependencies]
python = "^3.8"
torch = "2.0.0"
torchvision = "0.15.1"
tqdm = "4.65.0"
efficientnet_pytorch = "0.7.1"
numpy = "^1.18"
Pillow = "9.5.0"
pretrainedmodels = "0.7.4"
pydicom = "2.3.1"
requests = "2.28.2"
timm = "0.6.12"
scikit-image = "0.20.0"
matplotlib = "3.7.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
